requirements.txt
Flask==2.0.1
tensorflow==2.6.0
numpy==1.19.5
Pillow==8.3.1
pytest==6.2.4
pytest-flask==1.2.0